requirements.txt
streamlit
pandas
requests
deep_translator
python-docx
openpyxl
python-pptx
PyPDF2